[CL-1941] Move RenderNode to shared folder
