+++ front/app/components/ContentBuilder/RenderNode/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  error: {
+    id: 'app.containers.admin.ContentBuilder.error',
+    defaultMessage: 'error',
+  },
+});
